cadastrar produto

+++ app/Http/Controllers/ProductController.php
@@ -5,36 +5,74 @@ namespace App\Http\Controllers;
 use App\Models\product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
+
 
 class ProductController extends Controller
 {
     public function index()
     {
+        if (isset($_GET['limit'])) {
+            $limit = $_GET['limit'];
+        } else {
+            $limit = 10;
+        }
 
-        $products= product::paginate(10);
+        if (isset($_GET['search'])) {
+            $products = product::where('name', 'like', '%' . $_GET['search'] . '%')
+                ->where('description', 'like', '%' . $_GET['search'] . '%')
+                ->paginate($limit);
+        } else {
 
-        for( $i = 0; $i < count($products); $i++){
+            $products = product::paginate($limit);
+        }
+
+        for ($i = 0; $i < count($products); $i++) {
             $products[$i]->image = Storage::disk('public')->url($products[$i]->image);
         }
 
         return $products;
-
-
     }
 
     public function show($id)
     {
 
-        $product =product::find($id);
+        $product = product::find($id);
 
-        if(is_null($product)) return response()->json('Product not found', 404);
+        if (is_null($product)) return response()->json('Product not found', 404);
 
         return $product;
     }
 
     public function store(Request $request)
     {
-        $product = product::create($request->all());
+
+        $validator = Validator::make($request->all(), [
+            'image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
+            'name' => 'required|max:50',
+            'description' => 'required|max:200',
+            'price' => 'required|decimal:2',
+            'expiration_date' => 'required|date|after:yesterday',
+            'id_category' => 'required|exists:categories,id'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
+        }
+        // $validatedData = $request->validate([
+
+        // ]);
+
+        $products = $request->except(['image']);
+
+        if ($request->hasFile('image')) { //Verificar se existe imagem no request 
+
+            $imagePath = $request->file('image')->store('image', 'public');
+
+            $products['image'] = $imagePath;
+        }
+
+        $product = product::create($products);
 
         return $product;
     }
@@ -42,31 +80,41 @@ class ProductController extends Controller
     public function update(Request $request, $id)
     {
 
-        $product = product::find($id); //Buscar Produto por ID
+        $validator = Validator::make($request->all(), [
+            'image' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
+            'name' => 'max:50',
+            'description' => 'max:200',
+            'price' => 'decimal:2',
+            'expiration_date' => 'date|after:yesterday',
+            'id_category' => 'exists:categories,id'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
+        }
+
+        $product = product::find($id);
+
+        if (is_null($product)) return response()->json('Product not found', 404);
 
 
-        if(is_null($product)) return response()->json('Product not found', 404);
-
-    
         if ($request->hasFile('image')) { //Verificar se existe imagem no request 
 
-            $validatedData = $request->validate([
-                'image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
-              
-            ]);
-            if ($product->image && Storage::disk('public')->exists($product->image)) { 
+
+            if ($product->image && Storage::disk('public')->exists($product->image)) {
                 Storage::disk('public')->delete($product->image);
             }
 
-            $imagePath = $request->file('image')->storeAs('image', 'public');
+
+            $imagePath = $request->file('image')->store('image', 'public');
 
             $product->update([
                 'image' => $imagePath
             ]);
-        
         }
 
-        $product->update($request->all()); 
+        $product->update($request->except(['image']));
+
 
         return product::find($id);
     }
@@ -75,9 +123,9 @@ class ProductController extends Controller
     {
         $product = product::find($id);
 
-        if(is_null($product)) return response()->json('Product not found', 404);
+        if (is_null($product)) return response()->json('Product not found', 404);
 
-        if ($product->image && Storage::disk('public')->exists($product->image)) { 
+        if ($product->image && Storage::disk('public')->exists($product->image)) {
             Storage::disk('public')->delete($product->image);
         }
